+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.32] - 2025-07-28
+
+## Added
 - Allow administrators to add app passwords.
 - DeSEC DNS support.
 - OVH DNS support.
-
-## Added
 
 ### Changed